moved parse test to utils
